drm/i2c: dummy: Add dummy HDMI device
ARM Versatile Express motherboards contain a Silicon Image SiI9022 HDMI
encoder which unfortunately can not be reliably accessed by the kernel.
To work around this we provide a dummy driver to provide the interfaces
DRM drivers might expect but which doesn't actually access the hardware
in any way. Setup of the HDMI encoder hardware must therefore be
achieved through other means. On Versatile Express this can be done with
the DVIMODE and MUXFPGA values in config.txt

This dummy driver was constructed by copying tda998x_drv.c and stripping
out code and renaming functions. It contains a hard coded EDID for a
1024x768 display (generated from Documentation/EDID/Makefile).

Signed-off-by: Jon Medhurst <tixy@linaro.org>