drm: hdlcd: Implement a wait for scanout change to be in effect
User-side code doing double-buffering expects this in the situations
that ends up using mode_set_base.

Also only set scanout as needed, not on every interrupt, as that isn't
required and is unsafe if the framebuffer state is part way through
being updated.

Code adapted from an earlier version of the driver which was written by
Liviu Dudau <Liviu.Dudau@arm.com>

Signed-off-by: Jon Medhurst <tixy@linaro.org>