Revert "ARM64: add option to build Image.gz/dtb combo"
This reverts commit 082d89f7f2ae529e2d7b50f6259e4430472cf4ed because it
causes errors with 'make clean' when merged into kernels with dts files
in vendor subdirs i.e. LSK 3.18 which has backported commit
ca5b34100c571. We revert rather than fix this commit because we don't
need the feature and it's the sort of thing that wouldn't be accepted in
the mainline kernel.

Conflicts:
	arch/arm64/Kconfig
	arch/arm64/boot/Makefile
	arch/arm64/boot/dts/Makefile